moved the right card navbar to its own component. Also moved the right card components to its own folder
