+++ components/right-card/RightCard.tsx
@@ -1,0 +1,12 @@
+import RightCardNavBar from "./RightCardNavBar";
+
+function RightCard() {
+  return (
+    <div className="w-3/5 overflow-auto mt-2 flex flex-row">
+      <div className="w-full">hello</div>
+      <RightCardNavBar />
+    </div>
+  );
+}
+
+export default RightCard;
